consolidated hastext into hawc
