+++ hawc/apps/epiv2/templates/epiv2/fragments/_design_table.html
@@ -1,5 +1,3 @@
-{% load hastext %}
-
 <table class="table table-sm table-striped" hx-target="this" id="design-table">
     <colgroup>
         <col style="width: 20%;">
