+++ hawc/apps/study/templates/study/study_detail.html
@@ -1,7 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load hastext %}
-
 {% block content %}
   <div class="d-flex">
     <h2>{{object}}</h2>
